Add constants and move interfaces #20
