+++ user-manager/src/constants/index.ts
@@ -1,1 +1,2 @@
 export { VALIDATION_MESSAGE } from '@constants/messages';
+export { INFO_TYPE } from '@constants/types';
