+++ user-manager/src/constants/types.ts
@@ -1,0 +1,7 @@
+/**
+ * Represents types for the 'SideBarInfo' component.
+ */
+export const INFO_TYPE = {
+  TEXT_VIEW: 'textView',
+  LIST_VIEW: 'listView'
+};
