+++ user-manager/src/types/index.ts
@@ -2,3 +2,4 @@ export type { IPopoverOption } from './interfaces/popoverOption';
 export type { IItemNav } from './interfaces/itemNav';
 export type { IColumnProps } from './interfaces/columns';
 export type { IUserProps } from './interfaces/users';
+export type { ItemViewProps } from './interfaces/itemView';
