+++ user-manager/src/types/interfaces/itemView.ts
@@ -1,0 +1,8 @@
+export interface ItemViewProps {
+  icon: string;
+  title: string;
+  content: Array<{
+    text: string;
+    link: string;
+  }>;
+}
